[POC] collect the class into namespace, but seems cannot be generated by localization tool.
