+++ osu.Game.Rulesets.Karaoke/Edit/Generator/TimeTags/TimeTagGenerator.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Localisation;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Types;
 using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Rulesets.Karaoke.Localisation;
+using osu.Game.Rulesets.Karaoke.Localisation.Edit.Generator;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Generator.TimeTags
 {
