+++ osu.Game.Rulesets.Karaoke/Localisation/Edit/Generator/TimeTagGeneratorStrings.cs
@@ -1,10 +1,10 @@
 ﻿using osu.Framework.Localisation;
 
-namespace osu.Game.Rulesets.Karaoke.Localisation
+namespace osu.Game.Rulesets.Karaoke.Localisation.Edit.Generator
 {
-    public static class RomajiTagGeneratorStrings
+    public static class TimeTagGeneratorStrings
     {
-        private const string prefix = @"osu.Game.Rulesets.Karaoke.Localisation.RomajiTagGenerator";
+        private const string prefix = @"osu.Game.Rulesets.Karaoke.Localisation.TimeTagGenerator";
 
         /// <summary>
         /// "Lyric should not be empty."
